Custom_tkinter_V2/bibliotheque_idsi/requirements.txt
# ============================================
# Système de Gestion de Bibliothèque - IDSI
# Requirements Python
# ============================================

# Base de données
SQLAlchemy>=2.0.0

# Interface graphique moderne
customtkinter>=5.2.0

# Machine Learning pour les recommandations
numpy>=1.24.0
scikit-learn>=1.3.0

# Traitement d'images
pillow>=10.0.0

# Pour les graphiques statistiques (optionnel)
matplotlib>=3.7.0

# Export Excel (optionnel)
openpyxl>=3.1.0
